+++ packages/@leight/react-cmdk/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./store";
 export * from "./components/CommandPalette";
-export * from "./utils/utils";
-export * from "./utils/withItem";
+export * from "./utils";
 export * from "./types";
 
 import CommandPalette from "./components/CommandPalette";