+++ packages/@leight/react-cmdk/src/store/index.ts
@@ -1,1 +1,1 @@
-export * from "./useCommandPaletteStore";
+export * from "./useCommandPaletteState";