+++ packages/@leight/react-cmdk/src/utils/context.ts
@@ -11,20 +11,12 @@ export const SelectContext = createContext<{ selected: number }>({
 export const PageContext = createContext<{
     setSearchPrefix?: Dispatch<SetStateAction<string[] | undefined>>;
     searchPrefix?: string[];
-    page?: string;
 }>({
     searchPrefix: undefined,
-    page:         undefined,
-});
-
-export const SearchContext = createContext<{ search: string }>({
-    search: "",
 });
 
 export const OpenContext = createContext<{
     onChangeOpen: (value: boolean) => void;
-    isOpen: boolean;
 }>({
     onChangeOpen: () => undefined,
-    isOpen:       false,
 });