+++ packages/@leight/utils-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/utils-client
 
+## 0.5.19
+
+### Patch Changes
+
+- Improved zustand store stuff
+- Updated dependencies
+    - @leight/context-client@0.5.20
+
 ## 0.5.18
 
 ### Patch Changes